+++ docs/changes.rst
@@ -49,6 +49,7 @@ Release history
 - feat: selective forward settings API for workspaces: :attr:`api.workspace_settings.selective_forward <wxc_sdk.workspace_settings.WorkspaceSettingsApi.selective_forward>`
 - fix: Paging.routing_prefix instead of .routingPrefix
 - feat: new attribute AutoTransferNumbers.use_custom_transfer_numbers
+- feat: new attribute CallRecordingSetting.call_recording_access_settings
 
 1.20.0
 ------